--- web/src/components/users/user-details.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-function UserDetails() {
-  return (
-    <div>
-      <h1>
-        User Details For A Single User/Also re-used by Password Reset or Edit
-        Forms
-      </h1>
-    </div>
-  );
-}
-
-export default UserDetails;
